+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## Unreleased
+
+### Bugfixes
+
+- Fixed `GicV3::setup` not to write to GICD IGROUPR[0].
+
+### Improvements
+
+- Added more interrupt types to `IntId`, and public constants for number of each type.
+- Added constants to `IntId` for special interrupt IDs.
+
+## 0.1.2
+
+### Bugfixes
+
+- Changed `irouter` and `irouter_e` fields of `GICD` to use u64, to match GIC specification.
+
+### Improvements
+
+- Made `gicv3::registers` module public and added methods to `GicV3` to get pointers to registers.
+
 ## 0.1.1
 
 ### Improvements